requirements.txt
cython
grpcio
grpcio-tools
huggingface-hub==0.23.5
consulate